+++ mimesis/providers/person.py
@@ -304,6 +304,17 @@ class Person(BaseDataProvider):
 
         return self.random.choice(self._data['gender'])
 
+    def sex(self, *args, **kwargs):
+        """An alias for method self.gender()
+
+        See docstrings of method self.gender() for details.
+
+        :param args: Positional arguments.
+        :param kwargs: Keyword arguments.
+        :return: Sex
+        """
+        return self.gender(*args, **kwargs)
+
     def height(self, minimum: float = 1.5, maximum: float = 2.0) -> str:
         """Generate a random height in M (Meter).